+++ NEWS.md
@@ -1,5 +1,6 @@
 # googleCloudRunner 0.2.0.9000
 
+* Move the setup wizard functions from `googleCloudRunner` to `googleAuthR` so they are available for all packages.
 * Check for Cloud Scheduler Service Agent is present for scheduler to work (#73)
 
 # googleCloudRunner 0.2.0